refactor(cli): split augment_cmd into a package
augment_cmd had grown to ~1320 lines covering four distinct hook
concerns behind one flat module: search (Grep/Glob) enrichment,
per-session read/edit intelligence, post-commit stale-wiki detection,
and Codex lifecycle events, plus the Click entry and dispatch glue.

Split it into an augment_cmd/ package:

- _shared.py         repo-root / output-text / relativize leaf helpers
- search.py          Grep/Glob rescue, triage and flood-digest
- read_state.py      per-session read/edit tracking and skeleton nudges
- bash_staleness.py  post git-commit stale-wiki warnings
- codex.py           Codex context and post-edit handlers
- command.py         Click entry, stdin dispatch, emission dedup

Pure structural move: symbols relocate verbatim and all heavy imports
stay function-local, so the augment_hook fast path is unaffected.
__init__.py re-exports the public surface, so every importer (main.py,
augment_hook.py, tests) is unchanged. The _search_enrich monkeypatch
targets in test_augment_search.py are repointed to augment_cmd.search
to match the helper's new home (patch address only, no test logic
changed).
